File/peoples-bill-ghana/peoples-bill-ghana/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# ML & Data Processing
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
sentence-transformers==2.2.2  # For production clustering

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# SMS Integration (Phase 2)
# africastalking==1.2.5  # Uncomment when ready

# Optional: For OpenAI integration
# openai==1.3.7  # Uncomment if using GPT for clause generation
